Extract `DateTimePrecisionTest`
The datetime precision tests for any adapters is duplicated.
